requirements.txt
pytest
requests
faker
python-dotenv
allure-pytest